pyproject.toml
[tool.poetry]
name = "slack-integration"
version = "0.1.0"
description = "Slack integration using OAUTH2"
authors = ["ridwanray <alabarise@gmail.com>"]
readme = "README.md"
#packages = [{include = "slack_integration"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
python-decouple = "^3.8"
requests = "^2.31.0"
pydantic-settings = "^2.0.3"
httpx = "^0.25.0"
aiohttp = "^3.8.5"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-mock = "^3.11.1"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
